+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Pointer event coordinates fixed when canvas is in CSS transformed view.
 - Title appearance and disappearance glitch is fixed.
 - Anim.Control.seek() won't fail when Number passed instead of string.
+- When animation canceled, promise rejected with proper custom error. 
 
 ### Added